Fixed errors in rsakpMake introduced in version 3.0.0; omitted to adapt function to new parameter 'size in bits' style. Keypair generation function now employs new mpprndr_w function to generate the second prime in defined interval, following method stipulated by IEEE P.1363.

+++ rsakp.h
@@ -77,7 +77,7 @@ extern "C" {
 #endif
 
 BEECRYPTAPI
-int rsakpMake(rsakp*, randomGeneratorContext*, int);
+int rsakpMake(rsakp*, randomGeneratorContext*, size_t);
 BEECRYPTAPI
 int rsakpInit(rsakp*);
 BEECRYPTAPI
